PM / devfreq: bw_hwmon: Fix AB vote less than measurement when freq at max
When the decision window is terminated early by an IRQ and the frequency
voted by the governor is the max_freq allowed for that device, the AB vote
can be lower than the measured value. This really shouldn't affect power or
performance since the device frequency is already maxed out, but fix it
anyway so that the logs don't look suspicious.

Signed-off-by: Saravana Kannan <skannan@codeaurora.org>
Change-Id: I1788cbb1af026f1c54f9acc4806cb29cd6069e04
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>