pstore: drop pmsg bounce buffer
Removing a bounce buffer copy operation in the path is always
better, but we also gain on performance by not requesting a vmalloc
on every write as this can cause precious RT tasks to stall while
memory is being arranged for. We did this with a write_buf handler
signature change with an enum pstore_from_id argument as part of the
buffer specification passed into pstore ram core.

Pedantically added error propagation support to the write path in the
ram core, despite how impossible it is for an error could ever occur.

Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Bug: 31057326
Change-Id: I4cdee1cd31467aa3e6c605bce2fbd4de5b0f8caa
Signed-off-by: mydongistiny <jaysonedson@gmail.com>